trying to remove the big private packages
