--- rats-devtools/src/python/rats/_examples/__init__.py
@@ -1,5 +1,0 @@
-from ._plugin import ExamplesPluginContainer
-
-__all__ = [
-    "ExamplesPluginContainer",
-]
