+++ rats-devtools/src/python/rats/examples/__init__.py
@@ -1,0 +1,8 @@
+"""The beginnings of a library for easily registering examples."""
+
+from ._plugin import PluginContainer, PluginServices
+
+__all__ = [
+    "PluginContainer",
+    "PluginServices",
+]
