+++ rats-devtools/src/python/rats/examples/_cli.py
@@ -1,0 +1,11 @@
+from rats import cli
+
+
+class PluginCommands(cli.CommandContainer):
+    @cli.command()
+    def _ping(self) -> None:
+        print("ping")
+
+    @cli.command()
+    def _pong(self) -> None:
+        print("pong")
